requirements.dev.txt
django==3.1.6
pillow==8.1.2
social-auth-app-django==3.4.0
requests==2.25.1
psycopg2-binary
django-debug-toolbar
django-debug-toolbar-template-profiler
django-extensions
django-redis
